🔒️ 権限の調整
